method 1/requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
pymongo>=4.6.0
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
litellm>=1.0.0
